Added index page to fix middleware
