+++ src/middleware.ts
@@ -1,16 +1,13 @@
-import { authMiddleware } from "@clerk/nextjs/server";
+import { clerkMiddleware, createRouteMatcher } from "@clerk/nextjs/server";
 
-export default authMiddleware({
-    publicRoutes: ["/api/webhook", "/submit"],
-    ignoredRoutes: [
-        "/api/submissions/get",
-        "/api/submissions/add",
-        "/api/submissions/get-by-user",
-        "/api/submissions/edit",
-        "/api/submissions/delete"
-    ]
-});
+const isProtectedRoute = createRouteMatcher(
+    ["/home", "/submit"]
+);
 
-export const config = {
-    matcher: ["/((?!.+\\.[\\w]+$|_next).*)", "/", "/(api|trpc)/(.*)"]
-};
+export default clerkMiddleware((auth, req) => {
+    if (isProtectedRoute(req)) auth().protect();
+  });
+  
+  export const config = {
+    matcher: ['/((?!.*\\..*|_next).*)', '/', '/(api|trpc)(.*)'],
+  };
