+++ src/styles/globals.css
@@ -102,7 +102,7 @@ body {
     height: auto; 
 }
   
-UserEdit-image {
+.UserEdit-image {
     width: 100%;
     height: auto;
 }
